% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_fpr}
\alias{get_fpr}
\title{KP: What is the point of these functions when these values have already been computed in the above get_confusion_matrix function?
Calculates the False Positive Rate (FPR)}
\usage{
get_fpr(predictions, labels, threshold)
}
\arguments{
\item{predictions}{list of numerics,  predicted values}

\item{labels}{list of numerics, actual values/outcomes}

\item{threshold}{numeric, value between 0 - 1 to cut  predictions that are continous within binary 0s and 1s}
}
\value{
numeric, false positive rate
}
\description{
This function takes the predictions of a model,  (can be either binary 0 or 1, or continous numeric [0,1]) and
calculates the False Positive Rate.  Given that predictions need to be binary for the FPR to be calculated
you need to pass in a threshold to cut the predictions off.  If the predictions are already binary, then pass in .5
FPR = FP / (FP + TN)
}
