% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_tpr}
\alias{get_tpr}
\title{Calculates the True Positive Rate (TPR)}
\usage{
get_tpr(predictions, labels, threshold)
}
\arguments{
\item{predictions}{list of numerics,  predicted values}

\item{labels}{list of numerics, actual values/outcomes}

\item{threshold}{numeric, value between 0 - 1 to cut  predictions that are continous within binary 0s and 1s}
}
\value{
numeric, true positive rate
}
\description{
This function takes the predictions of a model,  (can be either binary 0 or 1, or continous numeric [0,1]) and
calculates the True Positive Rate.  Given that predictions need to be binary for the TPR to be calculated
you need to pass in a threshold to cut the predictions off.  If the predictions are already binary, then pass in .5
TPR = TP / (FN + TP)
}
