% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearInterpolate.R
\name{LinInterpolate}
\alias{LinInterpolate}
\title{2-D plane Linear Interpolation
A linear interpolation function using point slope formula to return a single point.}
\usage{
LinInterpolate(y0, y1, x0, x1, x)
}
\arguments{
\item{y0}{Initial point on the y-axis on a 2-D plane}

\item{y1}{End point on the y-axis on a 2-D plane}

\item{x0}{Initial point on the x-axis on a 2-D plane}

\item{x1}{End point on the y-axis on a 2-D plane}

\item{x}{Known point on the x-axis on a 2-D plane}
}
\value{
y Missing point on the y-axis on a 2-D plane
}
\description{
2-D plane Linear Interpolation
A linear interpolation function using point slope formula to return a single point.
}
\examples{
LinInterpolate(3,5,1,3,2)
}
