% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictiveperformancemetrics.R
\name{recall}
\alias{recall}
\title{Calculate the Recall for predictions}
\usage{
recall(predictions, outcomes, threshold)
}
\arguments{
\item{predictions}{list of numerics,  predicted values}

\item{outcomes}{list of numerics, actual values/outcomes}

\item{threshold}{numeric, value between 0 - 1 to cut  predictions that are continous within binary 0s and 1s}
}
\value{
numeric, returns recall value
}
\description{
This function takes the predictions of a model, (can be either binary 0 or 1, or continous numeric [0,1]) and
calculates the recall based on the predictions.  Given that predictions need to binary for the recall calculates
you need to pass in a threshold to but the predictions off.  If the predictions are already binary, then pass in .5
}
