% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{threshold_finder}
\alias{threshold_finder}
\title{Finds the threshold to cut continous predictions to maximize a given metric}
\usage{
threshold_finder(func, predictions, labels)
}
\arguments{
\item{func}{function, function that takes arguments, predictions, labels, and threshold}

\item{predictions}{list of numerics,  predicted values}

\item{labels}{list of numerics, actual values/outcomes}
}
\value{
numeric, value from 0 to 1 that is the threshold that maximizes the metric func
}
\description{
This function takes a given performance metric (precision, accuracy, or recall) that requires a threshold
and returns the threshold value that maximizes the specified metric
}
