% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_auc_roc}
\alias{get_auc_roc}
\title{Calculates AUC ROC for continious predictions, and actual predictions}
\usage{
get_auc_roc(predictions, labels)
}
\arguments{
\item{predictions}{list of numerics,  predicted values}

\item{labels}{list of numerics, actual values/outcomes}
}
\value{
numeric, returns AUC ROC value
}
\description{
This function takes the predictions of a model, (can be either binary 0 or 1, or continous numeric [0,1]) and
calculates the roc curve, and then gets the auc under the curve given the predicted values and actual values
Notes : baseline is .5, and perfect is 1
}
