% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_confusion_matrix}
\alias{get_confusion_matrix}
\title{Calculates Confusion Matrix (TP, FP, FN, TN)}
\usage{
get_confusion_matrix(predictions, labels, threshold)
}
\arguments{
\item{predictions}{list of numerics,  predicted values}

\item{labels}{list of numerics, actual values/outcomes}

\item{threshold}{numeric, value between 0 - 1 to cut  predictions that are continous within binary 0s and 1s}
}
\value{
table, table of integers with FP, FN, TP,R
}
\description{
This function takes the predictions of a model,  (can be either binary 0 or 1, or continous numeric [0,1]) and
calculates the confusion metrics based on the predictions.  Given that predictions need to be binary for the confusion matrix to work
you need to pass in a threshold to cut the predictions off.  If the predictions are already binary, then pass in .5
}
