% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictiveperformancemetrics.R
\name{threshold_finder}
\alias{threshold_finder}
\title{Finds the threshold to cut continous predictions to maximize a given metric}
\usage{
threshold_finder(func, predictions, outcomes)
}
\arguments{
\item{func}{function, takes in other functions as arguments such as precision, accuracy and recall}

\item{predictions}{list of numerics,  predicted values}

\item{outcomes}{list of numerics, actual values/outcomes}
}
\value{
numeric, value from 0 to 1 that is the threshold that maximizes the metric func
}
\description{
This function takes a given performance metric (precision, accuracy, or recall) that requires a threshold
and returns the threshold value that maximizes the specified metric
}
